connect to database

+++ model/user_model.go
@@ -1,7 +1,7 @@
 package model
 
 type User struct{
-	Id uint32 `json:"id" gorm:"type:int unsigned;primary_key;unique_index:id_UNIQUE;not null;"`
+	Id int64`json:"id" gorm:"type:int unsigned;primary_key;unique_index:id_UNIQUE;not null;"`
 	Name string `json:"login" gorm:"type:varchar(50);unique_index:name_UNIQUE;not null"`
 	DisplayName string `json:"name" gorm:"type:varchar(50);default:null"`
 	Email string `json:"email" gorm:"type:varchar(50);unique_index:email_UNIQUE;not null"`
